// src/types/shared.ts
export type Option = {
    value: string
    label: string
}

export enum ProductStatus {
    DRAFT = 'draft',
    PROPOSED = 'proposed',
    PUBLISHED = 'published',
    REJECTED = 'rejected',
}

export type DateFilter = null | {
    gt?: string
    lt?: string
}

export enum TaxRateType {
    REGION = 'region',
    RATE = 'rate',
}

export type PaginationProps = {
    limit: number
    offset: number
}

export type Idable = { id: string; [x: string]: any }